requirements.txt
Flask
openai # Görevleri bölmek ve agentları çağırmak için LLM kullanabiliriz
python-dotenv
requests # Diğer servislere HTTP çağrıları yapmak için
gunicorn
Flask-Cors # Frontend entegrasyonu için (ileride gerekebilir)
